Updated/improved gather-browser-info implementation.
