+++ wicket-core/src/main/java/org/apache/wicket/markup/html/pages/AbstractBrowserInfoForm.html
@@ -35,6 +35,7 @@
 			<input type="text" wicket:id="browserWidth" />
 			<input type="text" wicket:id="browserHeight" />
 			<input type="text" wicket:id="hostname" />
+			<wicket:child />
 		</form>
 	</wicket:panel>
 </html>
