+++ wicket-core/src/main/java/org/apache/wicket/markup/html/pages/AbstractBrowserInfoPage.html
@@ -15,7 +15,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 -->
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!DOCTYPE html>
 <html xmlns:wicket="http://wicket.apache.org">
 	<head>
 		<meta wicket:id="meta" http-equiv="refresh" />
@@ -24,6 +24,7 @@
 		If you see this, it means that both javascript and meta-refresh are not support by
 		your browser configuration. Please click <a wicket:id="link" href="#">this link</a> to
 		continue to the original destination.
+		<wicket:child />
 		<div wicket:id="postback"></div>
 	</body>
 </html>
