+++ wicket-core/src/main/java/org/apache/wicket/markup/html/pages/BrowserInfoPage.java
@@ -16,20 +16,9 @@
  */
 package org.apache.wicket.markup.html.pages;
 
-import org.apache.wicket.AttributeModifier;
-import org.apache.wicket.markup.head.IHeaderResponse;
-import org.apache.wicket.markup.head.OnLoadHeaderItem;
-import org.apache.wicket.markup.html.WebComponent;
-import org.apache.wicket.markup.html.WebMarkupContainer;
-import org.apache.wicket.markup.html.WebPage;
-import org.apache.wicket.model.AbstractReadOnlyModel;
-import org.apache.wicket.model.IModel;
-import org.apache.wicket.model.LoadableDetachableModel;
-import org.apache.wicket.protocol.http.ClientProperties;
-import org.apache.wicket.protocol.http.WebSession;
 import org.apache.wicket.protocol.http.request.WebClientInfo;
-import org.apache.wicket.request.cycle.RequestCycle;
 import org.apache.wicket.settings.IRequestCycleSettings;
+
 
 /**
  * <p>
@@ -46,97 +35,13 @@ import org.apache.wicket.settings.IRequestCycleSettings;
  * 
  * @author Eelco Hillenius
  */
-public class BrowserInfoPage extends WebPage
+public class BrowserInfoPage extends AbstractBrowserInfoPage<WebClientInfo, BrowserInfoForm>
 {
-	private static final long serialVersionUID = 1L;
-
-	private BrowserInfoForm browserInfoForm;
-
 	/**
-	 * Bookmarkable constructor. This is not for normal framework client use. It will be called
-	 * whenever JavaScript is not supported, and the browser info page's meta refresh fires to this
-	 * page. Prior to this, the other constructor should already have been called.
+	 * {@inheritDoc}
 	 */
-	public BrowserInfoPage()
-	{
-		initComps();
-		RequestCycle requestCycle = getRequestCycle();
-		WebSession session = (WebSession)getSession();
-		WebClientInfo clientInfo = session.getClientInfo();
-		if (clientInfo == null)
-		{
-			clientInfo = new WebClientInfo(requestCycle);
-			getSession().setClientInfo(clientInfo);
-		}
-		else
-		{
-			ClientProperties properties = clientInfo.getProperties();
-			properties.setJavaEnabled(false);
-		}
-
-		continueToOriginalDestination();
-	}
-
 	@Override
-	public void renderHead(IHeaderResponse response)
-	{
-		super.renderHead(response);
-
-		response.render(OnLoadHeaderItem.forScript(
-				String.format("Wicket.BrowserInfo.submitForm('%s')", browserInfoForm.getFormMarkupId())));
-	}
-
-	@Override
-	public boolean isVersioned()
-	{
-		return false;
-	}
-
-	/**
-	 * Adds components.
-	 */
-	private void initComps()
-	{
-		WebComponent meta = new WebComponent("meta");
-
-		final IModel<String> urlModel = new LoadableDetachableModel<String>()
-		{
-			private static final long serialVersionUID = 1L;
-
-			@Override
-			protected String load()
-			{
-				CharSequence url = urlFor(BrowserInfoPage.class, null);
-				return url.toString();
-			}
-		};
-
-		meta.add(AttributeModifier.replace("content", new AbstractReadOnlyModel<String>()
-		{
-			private static final long serialVersionUID = 1L;
-
-			@Override
-			public String getObject()
-			{
-				return "0; url=" + urlModel.getObject();
-			}
-
-		}));
-		add(meta);
-		WebMarkupContainer link = new WebMarkupContainer("link");
-		link.add(AttributeModifier.replace("href", urlModel));
-		add(link);
-
-		browserInfoForm = new BrowserInfoForm("postback")
-		{
-			private static final long serialVersionUID = 1L;
-
-			@Override
-			protected void afterSubmit()
-			{
-				continueToOriginalDestination();
-			}
-		};
-		add(browserInfoForm);
+	protected BrowserInfoForm createBrowserInfoForm(String formMarkupId) {
+		return new BrowserInfoForm(formMarkupId);
 	}
 }
